Improved README to mention important detail.
I think that sentence is not really true:
> Finally, invoke Run, which blocks until the first actor returns. 

We wanted to use this awesome package as recommended way of handling concurrent workers, but actually someone was saying that `oklog/run` is not waiting until all go routines exit. 0.o I knew it is not leaking any go routines, but actually the README caused that confusion. I guess it can mislead.

Comment for `Run` method actually states things properly :+1: