Deregister signal channel (#16)
If `SignalHandler` is being used as a mechanism to exit the program
entirely, these resources will be cleaned up naturally. But if not,
the signal should be explicitly deregistered.

Make sure the channel is eligible to be garbage collected and that no
additional signals should be delivered to the channel after this actor
exits.